+++ sktransf/scaler/__init__.py
@@ -1,0 +1,7 @@
+"""
+Scaler module
+"""
+
+from .standard import StandardScaler
+
+__all__ = ["StandardScaler"]