+++ sktransf/selector/__init__.py
@@ -1,0 +1,8 @@
+"""
+Selector module
+"""
+
+from .sku import DropSkuColumnSelector
+from .unique import DropUniqueColumnSelector
+
+__all__ = ["DropSkuColumnSelector", "DropUniqueColumnSelector"]