+++ sktransf/transformer/__init__.py
@@ -1,0 +1,8 @@
+"""
+Tranformer module
+"""
+
+from .bool import BoolColumnTransformer
+from .log import LogColumnTransformer
+
+__all__ = ["LogColumnTransformer", "BoolColumnTransformer"]