+++ docs/docs/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
+## [0.3.0] - 2024-02-10
+
+### Added
+- Add validators for all classes
+- Add Cleaner module 
+- Add StandardScaler implementation
+- Manage problem of data drift train / test 
+
+
+### Updated
+- Moove .utils to root folder
+- Reshape entire module using transformer, cleaner, scaler, and selector
+- Reshape pyproject and documentation
+
 ## [0.2.1] - 2024-02-09
 
 ### Added